luu lai

+++ controller/paymentController.js
@@ -20,6 +20,7 @@ const TicketFood = require("../model/ticketFoodModel");
 const Booking = require("../model/bookingModel");
 const User = require("../model/userModel");
 const PaymentMethod = require("../model/paymentMethodModel");
+const VoucherUse = require("../model/voucherUseModel");
 const Factory = require("./handleFactory");
 const Email = require("../utils/email");
 
@@ -81,14 +82,34 @@ const updateTicketStatus = async (orderId, status) => {
         })
       await sendBookingConfirmationEmail(ticket, seatIds, ticketFoods);
     }
-    else {
+    else if ( status === "Failed" ) {
       await Booking.deleteMany({
         seatId: { $in: seatIds },
         userId: ticket.userId,
         showtimeId: ticket.showtimeId,
       });
-    }
 
+      // Giảm lượt sử dụng lại
+      if(ticket.voucherUseId){
+          await VoucherUse.updateOne(
+              {
+                  _id: ticket.voucherUseId,
+                  usageCount: { $gt: 0 }
+              },
+              {
+                  $inc: { usageCount: -1 }
+              }
+          )
+      }
+
+      // **Tăng lại stock food**
+      for (const tf of ticketFoods) {
+        await Food.updateOne(
+            { _id: tf.foodId },
+            { $inc: { inventoryCount: tf.quantity } }
+        );
+      }
+    }
 }
 
 const sendBookingConfirmationEmail = async (ticket, seatIds, ticketFoods) => {
@@ -193,6 +214,7 @@ exports.createPayment = catchAsync(async (req, res, next) => {
 });
 
 exports.momoCallback = catchAsync(async (req, res, next) => {
+  console.log("CallBack từ momo");
   const valid = await verifyCallbackByGateway("momo", req.body);
 
   if (!valid.isValid) {
@@ -278,30 +300,21 @@ exports.queryMomoPayment = async (orderId) => {
   };
   try {
     const response = await axios(options);
+    const data  = response.data;
 
-    if (response.data.resultCode === 0) {
-      const { orderId } = response.data;
+    if (data.resultCode === 0) {
       await updateTicketStatus(orderId, "Paid");
     }else{
         // Cập nhật trạng thái thanh toán không thành công
         await updateTicketStatus(orderId, "Failed");
-        return next(new AppError("Thanh toán không thành công", 400));
     }
-
-    res.status(200).json({
-      status: "success",
-      data: {
-        resultCode: response.data.resultCode,
-        message: response.data.message,
-      },
-    });
   }catch (error) {
-    return next(new AppError(`Lỗi khi truy vấn MoMo: ${error.message}`, 500));
+      console.error(`[Momo] 🔥 Lỗi truy vấn thanh toán: ${error.message}`);
   }
 
 };
 
-exports.queryZaloPayPayment = async (appTransId) => {
+exports.queryZaloPayPayment = async (appTransId, orderId) => {
   let postData = {
     app_id: process.env.ZALOPAY_APP_ID,
     app_trans_id: appTransId, // Input your app_trans_id
@@ -321,26 +334,13 @@ exports.queryZaloPayPayment = async (appTransId) => {
 
   try{
     const result = await axios(postConfig);
-
     if (result.data.return_code === 1) {
-      const embedData = JSON.parse(result.data.embed_data);
-      const orderId = embedData.order_id;
-
       await updateTicketStatus(orderId, "Paid");
     } else if (result.data.return_code === 2) {
       await updateTicketStatus(orderId, "Failed");
-      return next(new AppError("Giao dịch thất bại", 400));
     }
-
-    res.status(200).json({
-        status: "success",
-        data: {
-            transactionStatus: result.data.return_code,
-            message: result.data.return_message
-        }
-    })
   }
   catch (error) {
-      return next(new AppError(`Lỗi khi truy vấn ZaloPay: ${error.message}`, 500));
+      console.error(`[ZaloPay] 🔥 Lỗi truy vấn thanh toán: ${error.message}`);
   }
 };
